MAINT: Rename calibration to target solely models.
