+++ src/darsia/multi_image_analysis/model_calibration.py
